refactor: Move internal types to separate files
