+++ src/components/sections/results/SpecialDurations.tsx
@@ -3,7 +3,7 @@ import { Table, TableBody, TableCell, TableRow, Typography } from '@mui/material
 import { Trans } from 'react-i18next';
 import { makeStyles } from 'tss-react/mui';
 import { boons, damagingConditions } from '../../../assets/modifierdata/metadata';
-import type { Attributes } from '../../../state/optimizer/optimizerCore';
+import type { Attributes } from '../../../state/optimizer/types/optimizerTypes';
 
 // const boonDurations = boons.map((name) => `${name} Duration`);
