+++ src/state/optimizer/optimizer.ts
@@ -6,14 +6,14 @@ import { defaultJsThreads } from '../slices/controlsSlice';
 import type { ExtrasType } from '../slices/extras';
 import type { RootState } from '../store';
 import { iteratePartitionCount } from './combinatorics';
+import { characterLT, UPDATE_MS } from './optimizerCore';
+import { setupCombinations } from './optimizerSetup';
+import type { ExtrasCombinationEntry } from './types/optimizerSetupTypes';
 import type {
   Character,
   OptimizerCoreMinimalSettings,
   OptimizerCoreSettings,
-} from './optimizerCore';
-import { characterLT, UPDATE_MS } from './optimizerCore';
-import type { ExtrasCombinationEntry } from './optimizerSetup';
-import { setupCombinations } from './optimizerSetup';
+} from './types/optimizerTypes';
 
 export interface Combination extends ExtrasCombinationEntry {
   index: number;
