+++ src/state/optimizer/optimizerSetup.ts
@@ -3,7 +3,6 @@
 /* eslint-disable dot-notation */
 
 import type {
-  AmountData,
   AttributeKey,
   AttributePointMode,
   ConversionAfterBuffsDestinationKey,
@@ -16,7 +15,6 @@ import type {
   DamageMode,
   DamageValue,
   Percent,
-  Modifiers as YamlModifiers,
 } from '../../assets/modifierdata/metadata';
 import {
   allAttributeCoefficientKeys,
@@ -45,14 +43,11 @@ import {
   parseInfusionCount,
   parsePriority,
 } from '../../utils/usefulFunctions';
-import type { BossSlice } from '../slices/boss';
 import { getAttackRate, getMovementUptime } from '../slices/boss';
-import type { BuffsSlice } from '../slices/buffs';
 import { getBuffsModifiers } from '../slices/buffs';
 import { getProfession } from '../slices/controlsSlice';
 import { getDistributionNew } from '../slices/distribution';
 import { getExtraModifiersModifiers } from '../slices/extraModifiers';
-import type { ExtrasCombination, ExtrasSlice } from '../slices/extras';
 import { getExtrasCombinationsAndModifiers, getExtrasIds } from '../slices/extras';
 import { getForcedSlots } from '../slices/forcedSlots';
 import {
@@ -60,7 +55,6 @@ import {
   getMaxInfusions,
   getValidInfusionOptions,
 } from '../slices/infusions';
-import type { PrioritiesSlice } from '../slices/priorities';
 import {
   getAffixes,
   getConstraint,
@@ -70,78 +64,34 @@ import {
   getOptimizeFor,
   getWeaponType,
 } from '../slices/priorities';
-import type { SkillsSlice } from '../slices/skills';
 import { getSkillsModifiers } from '../slices/skills';
-import type { TraitsSlice } from '../slices/traits';
 import { getCurrentSpecialization, getTraitsModifiers } from '../slices/traits';
 import { getGameMode } from '../slices/userSettings';
 import type { RootState } from '../store';
+import { clamp, scaleValue } from './optimizerCore';
+import type {
+  AppliedModifier,
+  DamageMultiplier,
+  DamageMultiplierBreakdown,
+  DistributionNameUI,
+  Modifiers,
+  MultiplierName,
+} from './types/optimizerSetupTypes';
 import type {
   OptimizerCoreSettings,
   OptimizerCoreSettingsPerCalculation,
   OptimizerCoreSettingsPerCombination,
-} from './optimizerCore';
-import { clamp, scaleValue } from './optimizerCore';
-
-export interface ExtrasCombinationEntry {
-  extrasCombination: ExtrasCombination;
-  extrasModifiers: AppliedModifier[];
-}
-
-type MultiplierName =
-  | 'Outgoing Strike Damage'
-  | 'Outgoing Condition Damage'
-  | 'Outgoing Siphon Damage'
-  | 'Incoming Strike Damage'
-  | 'Outgoing Critical Damage'
-  | 'Outgoing Bleeding Damage'
-  | 'Outgoing Burning Damage'
-  | 'Outgoing Confusion Damage'
-  | 'Outgoing Poison Damage'
-  | 'Outgoing Torment Damage'
-  | 'Outgoing Alternative Damage'
-  | 'Outgoing Alternative Critical Damage'
-  | 'Outgoing Phantasm Damage'
-  | 'Outgoing Phantasm Critical Damage';
-
-export interface AppliedModifier {
-  id?: string;
-  amount?: string;
-  modifiers: YamlModifiers;
-  wvwModifiers?: YamlModifiers;
-  amountData?: AmountData;
-  temporaryBuff?: true | false | 'activeOutOfCombat';
-}
-
-// todo: move these; they should be synchronized with ../../assets/modifierdata/metadata.js and
-// ../../components/sections/distribution/DamageDistribution.jsx
-// (unsure how that would best be done)
-export type DistributionNameUI =
-  | 'Power'
-  | 'Power2'
-  | 'Burning'
-  | 'Bleeding'
-  | 'Poisoned'
-  | 'Torment'
-  | 'Confusion';
-export type DistributionNameInternal =
-  | 'Power'
-  | 'Power2'
-  | 'Burning'
-  | 'Bleeding'
-  | 'Poison'
-  | 'Torment'
-  | 'Confusion';
+} from './types/optimizerTypes';
 
 type CollectedAttributeModifiers = Partial<Record<AttributeKey, number>>;
 
 type CollectedConversionValue = Partial<Record<ConversionSourceKey, number>>;
-export type CollectedConversionModifers = Partial<
+type CollectedConversionModifers = Partial<
   Record<ConversionDestinationKey, CollectedConversionValue>
 >;
 
 type CollectedConversionAfterBuffsValue = Partial<Record<ConversionAfterBuffsSourceKey, number>>;
-export type CollectedConversionAfterBuffsModifers = Partial<
+type CollectedConversionAfterBuffsModifers = Partial<
   Record<ConversionAfterBuffsDestinationKey, CollectedConversionAfterBuffsValue>
 >;
 
@@ -150,66 +100,6 @@ interface CollectedModifiers {
   convert: CollectedConversionModifers;
   convertAfterBuffs: CollectedConversionAfterBuffsModifers;
 }
-export interface MultiplierBreakdown {
-  mult: number;
-  add: number;
-  target: number;
-  total: number;
-}
-export type DamageMultiplier = Record<MultiplierName, number>;
-export type DamageMultiplierBreakdown = Record<MultiplierName, MultiplierBreakdown>;
-export interface Modifiers {
-  damageMultiplier: DamageMultiplier;
-  damageMultiplierBreakdown: DamageMultiplierBreakdown;
-  buff: [AttributeKey, number][];
-  convert: [ConversionDestinationKey, [ConversionSourceKey, number][]][];
-  convertAfterBuffs: [
-    ConversionAfterBuffsDestinationKey,
-    [ConversionAfterBuffsSourceKey, number][],
-  ][];
-}
-
-export type InfusionMode = 'None' | 'Primary' | 'Few' | 'Secondary' | 'SecondaryNoDuplicates';
-
-export interface CachedFormState {
-  traits: TraitsSlice;
-  skills: SkillsSlice;
-  extras: ExtrasSlice;
-  buffs: BuffsSlice;
-  priorities: PrioritiesSlice;
-  boss: BossSlice;
-}
-
-// interface OptimizerInput {
-//   profession: ProfessionName;
-//   specialization: string;
-//   weaponType: WeaponHandednessType;
-//   affixes: AffixName[]; // all selected gear affixes to iterate over
-//   forcedAffixes: (AffixName | null)[]; // array of specific affix names for each slot, or '' for unspecfied
-//   rankby: 'Damage' | 'Survivability' | 'Healing';
-//   minBoonDuration: number | null;
-//   minHealingPower: number | null;
-//   minToughness: number | null;
-//   maxToughness: number | null;
-//   minHealth: number | null;
-//   minCritChance: number | null;
-//   maxResults: number;
-//   primaryInfusion?: string;
-//   secondaryInfusion?: string;
-//   maxInfusions: number; // number of infusions, 0-18
-//   primaryMaxInfusions: number; // number of infusions, 0-18
-//   secondaryMaxInfusions: number; // number of infusions, 0-18
-//   distributionVersion?: 1 | 2; // version 1: old style (percentDistribution), verison 2: new style (coeff / sec)
-//   percentDistribution?: Record<string, number>; // old style distribution (sums to 100)
-//   distribution?: Record<string, number>; // new style distribution (coefficient * weaponstrength per second; average condition stacks)
-//   attackRate: number; // boss attack rate (for confusion)
-//   movementUptime: number; // boss movement uptime (for torment)
-
-//   appliedModifiers: AppliedModifier[]; // array of modifier objects
-
-//   infusionNoDuplicates: any;
-//   customAffixData: any;
-// }
 
 export function setupCombinations(reduxState: RootState) {
   console.groupCollapsed('Debug Info:');
